+++ src/main/java/com/mraof/minestuck/data/recipe/MinestuckGristCostsProvider.java
@@ -650,7 +650,14 @@ public class MinestuckGristCostsProvider extends RecipeProvider
 		GristCostRecipeBuilder.of(MSItems.CASSETTE_EMISSARY.get()).grist(BUILD, 15).grist(CAULK, 8).grist(GOLD, 5).grist(MARBLE, 5).buildFor(recipeSaver, Minestuck.MOD_ID);
 		GristCostRecipeBuilder.of(MSItems.CASSETTE_DANCE_STAB.get()).grist(BUILD, 15).grist(CAULK, 8).grist(COBALT, 5).grist(SULFUR, 5).buildFor(recipeSaver, Minestuck.MOD_ID);
 		GristCostRecipeBuilder.of(MSItems.CASSETTE_RETRO_BATTLE.get()).grist(BUILD, 15).grist(CAULK, 8).grist(QUARTZ, 5).grist(RUST, 5).buildFor(recipeSaver, Minestuck.MOD_ID);
+		GristCostRecipeBuilder.of(MSItems.CASSETTE_11.get()).grist(BUILD, 15).grist(CAULK, 8).grist(TAR, 5).grist(MERCURY, 5).buildFor(recipeSaver, Minestuck.MOD_ID);
+		GristCostRecipeBuilder.of(MSItems.CASSETTE_PIGSTEP.get()).grist(BUILD, 15).grist(CAULK, 8).grist(SULFUR, 5).grist(RUBY, 5).buildFor(recipeSaver, Minestuck.MOD_ID);
+		GristCostRecipeBuilder.of(MSItems.CASSETTE_STAL.get()).grist(BUILD, 15).grist(CAULK, 8).grist(TAR, 5).grist(RUST, 5).buildFor(recipeSaver, Minestuck.MOD_ID);
+		GristCostRecipeBuilder.of(MSItems.CASSETTE_STRAD.get()).grist(BUILD, 15).grist(CAULK, 8).grist(TAR, 5).grist(CHALK, 5).buildFor(recipeSaver, Minestuck.MOD_ID);
+		GristCostRecipeBuilder.of(MSItems.CASSETTE_WAIT.get()).grist(BUILD, 15).grist(CAULK, 8).grist(COBALT, 5).grist(DIAMOND, 5).buildFor(recipeSaver, Minestuck.MOD_ID);
+		GristCostRecipeBuilder.of(MSItems.CASSETTE_WARD.get()).grist(BUILD, 15).grist(CAULK, 8).grist(URANIUM, 5).grist(MERCURY, 5).buildFor(recipeSaver, Minestuck.MOD_ID);
 		GristCostRecipeBuilder.of(MSItems.CASSETTE_OTHERSIDE.get()).grist(BUILD, 60).grist(CAULK, 32).grist(URANIUM, 20).grist(COBALT, 20).buildFor(recipeSaver, Minestuck.MOD_ID);
+		GristCostRecipeBuilder.of(MSItems.CASSETTE_5.get()).grist(BUILD, 15).grist(CAULK, 8).grist(DIAMOND, 5).grist(COBALT, 5).buildFor(recipeSaver, Minestuck.MOD_ID);
 		
 		GristCostRecipeBuilder.of(MSBlocks.GLOWING_MUSHROOM.get()).grist(BUILD, 5).grist(SHALE, 3).grist(MERCURY, 2).build(recipeSaver);
 		GristCostRecipeBuilder.of(MSBlocks.GLOWING_MUSHROOM_VINES.get()).grist(BUILD, 8).grist(SHALE, 4).grist(MERCURY, 3).build(recipeSaver);